requirements.txt
mysql-connector-python
fastapi
uvicorn[standard]
pydantic
python-dotenv